refactor routes dir
